src/platform-bootstrap/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "platform-bootstrap"
version = "0.1.0"
description = "Dataset profiler and Supabase pgvector loader for Dataset Whisperer"
requires-python = ">=3.12"
dependencies = [
    "anthropic==0.40.0",
    "openai==1.57.4",
    "supabase==2.10.0",
    "google-cloud-storage==2.18.2",
    "pandas==2.2.3",
    "numpy==2.1.3",
    "pydantic==2.10.3",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "ruff==0.8.6",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
